Fix camera freezes on iOS by using a 0px video
On iOS, using a hidden `<video>` element causes the camera to freeze and turn off after a few seconds. It is not clear why this happens but might have to do with Safari's heuristics to disallow some autoplaying videos. Neither Safari's logs nor the iOS system logs explain the cause, but removing the `video.play()` call keeps the camera on, so the issue seems related to the video element.

Regardless of the cause, using a 0px video solves the issue. (Using a video element that's not part of the DOM tree should also work.)

Tested by running `yarn dev`, starting a network tunnel, and loading the URL from an iPhone and seeing that the camera stream works as expected. Also tested on desktop Chrome.