Fix tearDown routine
The effect hook in the Camera component called `tearDown()` when setting up instead of returning a callback to be invoked when the Camera switches which camera it uses.

With this, `setUp` doesn't need to call `tearDown` anymore since React will properly run the teardown callback before setting up the new camera.

Lastly, `tearDown` stopped all media streams and then the audio streams again, but it only needs to stop the video streams.

Tested by loading the site, giving it permission to access the camera, and verified the stream shows up as expected. Switched cameras via the dropdown menu and saw the viewfinder switch cameras. Switched back and saw the viewfinder update again.